- Create API to delete node edge and manage their group nodes - Create API to delete and update workflow and trigger - Create validations for the apis and implement error handling

+++ src/services/node-edge.service.ts
@@ -1,10 +1,13 @@
 import { NodeEdge } from "@prisma/client";
 import { prisma } from "../config";
 import { CreateNodeEdgeRecord, NodeEdgesCondition } from "../types";
+import { updateNodeParent } from "./node.service";
 
-export async function createNodeEdge(data: CreateNodeEdgeRecord): Promise<NodeEdge> {
+export async function createNodeEdge(data: CreateNodeEdgeRecord, updateParentGroup: boolean = true): Promise<NodeEdge> {
   try {
-    return await prisma.nodeEdge.create({ data });
+    const nodeEdge = await prisma.nodeEdge.create({ data });
+    updateParentGroup && data.group_id && (await updateNodeParent(data.target_node_id, data.group_id));
+    return nodeEdge;
   } catch (error) {
     console.error("ERROR: TO CREATE NODE EDGE", error);
     throw error;
@@ -31,6 +34,17 @@ export async function getNextNodeId(
   }
 }
 
+export async function getNodeEdge(id: string): Promise<NodeEdge> {
+  try {
+    const nodeEdge = await prisma.nodeEdge.findUnique({ where: { id } });
+    if (!nodeEdge) throw new Error("Edge Not Found");
+    return nodeEdge;
+  } catch (error) {
+    console.error("ERROR: TO GET NODE EDGE", error);
+    throw error;
+  }
+}
+
 export async function getNextNodeAfterLoop(loopNodeId: string): Promise<string | null> {
   try {
     const edges = await prisma.nodeEdge.findMany({
@@ -50,9 +64,18 @@ export async function getNextNodeAfterLoop(loopNodeId: string): Promise<string |
 
 export async function deleteNodeEdges(workflowId: string, sourceId: string, targetId: string): Promise<void> {
   try {
+    const nodeEdges = await prisma.nodeEdge.findMany({
+      where: { workflow_id: workflowId, source_node_id: sourceId, target_node_id: targetId },
+    });
+
+    if (!nodeEdges.length) return;
     await prisma.nodeEdge.deleteMany({
       where: { workflow_id: workflowId, source_node_id: sourceId, target_node_id: targetId },
     });
+
+    for (const edge of nodeEdges) {
+      edge.group_id && (await updateNodeParent(edge.target_node_id));
+    }
   } catch (error) {
     console.error("ERROR: TO DELETE NODE EDGES BETWEEN SOURCE AND TARGET", error);
     throw error;
@@ -61,9 +84,11 @@ export async function deleteNodeEdges(workflowId: string, sourceId: string, targ
 
 export async function deleteNodeEdgeById(edgeId: string): Promise<void> {
   try {
+    const nodeEdge = await getNodeEdge(edgeId);
     await prisma.nodeEdge.delete({
       where: { id: edgeId },
     });
+    nodeEdge.group_id && (await updateNodeParent(nodeEdge.target_node_id));
   } catch (error) {
     console.error("ERROR: TO DELETE NODE EDGE", error);
     throw error;
